+++ src/Pim/Bundle/CatalogBundle/MongoDB/Normalizer/MetricNormalizer.php
@@ -3,16 +3,16 @@
 namespace Pim\Bundle\CatalogBundle\MongoDB\Normalizer;
 
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
-use Pim\Bundle\CatalogBundle\Model\ProductPrice;
+use Pim\Bundle\CatalogBundle\Model\Metric;
 
 /**
- * Normalize a product price collection to store it as mongodb_json
+ * Normalize a product metric to store it as mongodb_json
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ProductPriceNormalizer implements NormalizerInterface
+class MetricNormalizer implements NormalizerInterface
 {
     /**
      * {@inheritdoc}
@@ -22,10 +22,11 @@ class ProductPriceNormalizer implements NormalizerInterface
         $data = null;
         if ($object->getData() !== null) {
             $data = [
-                $object->getCurrency() => [
-                    'data'     => $object->getData(),
-                    'currency' => $object->getCurrency()
-                ]
+                'data'     => $object->getData(),
+                'unit'     => $object->getUnit(),
+                'baseData' => $object->getBaseData(),
+                'baseUnit' => $object->getBaseUnit(),
+                'family'   => $object->getFamily()
             ];
         }
 
@@ -37,6 +38,6 @@ class ProductPriceNormalizer implements NormalizerInterface
      */
     public function supportsNormalization($data, $format = null)
     {
-        return $data instanceof ProductPrice && 'mongodb_json' === $format;
+        return $data instanceof Metric && 'mongodb_json' === $format;
     }
 }