+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Mass edit operator has been moved to an Operator sub-namespace
 - Pim\Bundle\EnrichBundle\MassEditAction\Operation\MassEditActionInterface has been renamed Pim\Bundle\EnrichBundle\MassEditAction\Operation\MassEditOperationInterface
 - Changed the HydratorInterface::hydrate() method signature
+- Avoid to store null values in Product::normalizedData (MongoDB support)
 
 # 1.1.0 - "Rabbit Punch" (2014-04-16)